Grouping all IntEnum subclasses to a new "enums" module
